Move around integration tests
