+++ vimwiki-core/tests/lib.rs
@@ -1,4 +1,2 @@
-mod fixtures;
-mod integration;
 mod output;
-mod utils;
+mod parser;
