+++ vimwiki-core/tests/parser/fixtures/vimwikiwiki/Tips and Snips.wiki
