+++ vimwiki-core/tests/parser/fixtures/vimwikiwiki/index.wiki
