+++ vimwiki-core/tests/parser/integration/misc/comment_in_definition_list.rs
@@ -1,4 +1,4 @@
-use crate::fixtures::VimwikiFile;
+use crate::parser::fixtures::VimwikiFile;
 use std::borrow::Cow;
 use vimwiki_core::*;
