+++ vimwiki-core/tests/parser/integration/misc/windows_support.rs
@@ -1,4 +1,4 @@
-use crate::fixtures::VimwikiFile;
+use crate::parser::fixtures::VimwikiFile;
 use std::{borrow::Cow, collections::HashMap};
 use vimwiki_core::*;
