+++ vimwiki-core/tests/parser/mod.rs
@@ -1,0 +1,3 @@
+mod fixtures;
+mod integration;
+mod utils;
